Linux port: sweep remaining small clang warning categories to zero
Per-site fixes across ~17 source files for warnings that didn't fall under
any of the prior bulk-fix categories. Notable items:

- TYPESAssertSz: replace `*((ULONG*)NULL) = 9;` with __builtin_trap()
  on clang/GCC. The store-through-NULL idiom is UB and clang elides it
  rather than trapping (Wnull-dereference, 288 sites). __builtin_trap
  generates a real ud2 / SIGILL.

- SzNewFile / UlNewLine: drop the INLINE attribute. The bodies live in
  os/memory.cxx; declaring them inline at the header but defining out
  of line trips Wundefined-inline (26 sites).

- dbshrink.cxx SHKPerformRootMove: two sites reference `fRecoveringRedo`
  (the integer-1 enum value of `enum RECOVERING_MODE`) where the
  function parameter `fRecoveryRedo` was almost certainly intended.
  Because `fRecoveringRedo` always evaluates to 1, the `&& fRecoveringRedo`
  gate at line ~2366 is a no-op and the `Assert( fRecoveringRedo )` at
  line ~2383 never fires — yielding behavior that has been latent since
  the code was written. Rather than silently change a long-standing
  runtime path, both sites are commented out with detailed NOTE blocks
  preserving the original semantics, pending clarification from
  Microsoft. Tracked separately for follow-up.

- LRD bitfield: change SHORT to USHORT so `{ ..., 1 }` fits the 1-bit
  fDebugOnly field (Wsingle-bit-bitfield-constant-conversion).

- BEGIN_PRAGMA_OPTIMIZE_DISABLE / END_PRAGMA_OPTIMIZE: gate the
  __pragma(optimize(...)) expansion to MSVC; clang ignores the
  optimize pragma (Wignored-pragmas).

Plus small cleanups: empty-body for-loops, vexing-parse extern,
trigraph escape, ULL suffix on a 64-bit FNV constant, INT_PTR
intermediate cast for void* -> INT, and similar.

After this sweep: zero clang warnings on the Linux build (the
remaining 10 build failures are pre-existing _wcsupr_s / private-
member access issues unrelated to warnings).

+++ dev/ese/src/ese/space.cxx
@@ -14117,7 +14117,7 @@ LOCAL ERR ErrSPIGetFsSe(
             const CPG cpgLogicalFileSize = pgnoPreLast + cpgDBReserved;
 
             //  cpg of database should be reasonable
-            Enforce( pgnoPreLast < 0x100000000 );
+            Enforce( (QWORD)pgnoPreLast < 0x100000000 );
 
             //  this includes the DB headers in this cpg size, not cpg of database.
             Enforce( g_rgfmp[ pfucb->ifmp ].CbOwnedFileSize() / g_cbPage < 0x100000000 );
